Second Comit
The Connect class launches the application in a new JVM when iotdev mode is selected. 
There are new options to list iot apps, show iot app details, show iot logical plan and kill an iot app.
The IotDev class instantiates the StramLocalCluster and follows the working of the local mode but, in a new JVM instance.